Fix/ios parsing (#22)
* chore: re-add test file for file upload

* chore: update ios podfile

* fix(types): parse filesize to int

* chore: add generated file

* fix(bridge): parse values correctly

* fix(types): parse doubles to ints

+++ packages/enmeshed_types/lib/src/contents/attribute_values/identity/birth_date.dart
@@ -12,9 +12,9 @@ class BirthDate extends IdentityAttributeValue {
   });
 
   factory BirthDate.fromJson(Map<String, dynamic> json) => BirthDate(
-        day: json['day'],
-        month: json['month'],
-        year: json['year'],
+        day: json['day'].toInt(),
+        month: json['month'].toInt(),
+        year: json['year'].toInt(),
       );
 
   @override
